+++ agat/data/__init__.py
@@ -5,6 +5,6 @@ Created on Tue Aug 15 11:41:20 2023
 @author: ZHANG Jun
 """
 
-from agat.data.data import *
-from agat.data.AtomicFeatures import get_atomic_features, get_atomic_feature_onehot
+from .data import *
+from .AtomicFeatures import get_atomic_features, get_atomic_feature_onehot