+++ agat/lib/__init__.py
@@ -5,7 +5,7 @@ Created on Tue Aug 15 11:41:20 2023
 @author: ZHANG Jun
 """
 
-from agat.lib.adsorbate_poscar import *
-from agat.lib.AgatException import *
-from agat.lib.file import *
-from agat.lib.GatLib import *
+from .adsorbate_poscar import *
+from .AgatException import *
+from .file import *
+from .GatLib import *