+++ agat/model/__init__.py
@@ -5,7 +5,7 @@ Created on Tue Aug 15 11:41:20 2023
 @author: ZHANG Jun
 """
 
-from agat.model.GatEnergyModel import *
-from agat.model.GatForceModel import *
-from agat.model.SingleGatLayer import *
-from agat.model.ModelFit import *
+from .GatEnergyModel import *
+from .GatForceModel import *
+from .SingleGatLayer import *
+from .ModelFit import *